split the code into two separate files
page. and data
