+++ systemvetardagen2025frontend/src/components/ProjectContributors/teamData.js
@@ -1,0 +1,154 @@
+export const teamData = {
+    "Web Development": {
+      head: {
+        name: "Jan Pakos",
+        email: "jan.p.pakos@gmail.com",
+        linkedin: "https://linkedin.com/in/janesmith",
+        github: "https://www.linkedin.com/in/jan-p-b065a2127"
+      },
+      viceHead: {
+        name: "Daniel Mansour",
+        email: "",
+        linkedin: "",
+        github: ""
+      },
+      teamMembers: [
+        {
+          name: "Test member 2",
+          email: "",
+          linkedin: "",
+          github: ""
+        },
+        {
+          name: "Testmember 2",
+          email: "alice@systemvetardagen.se",
+          github: "https://github.com/alicej"
+        }
+      ]
+    },
+    "Marketing": {
+      head: {
+        name: "Emma Wilson",
+        email: "emma@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/emmaw"
+      },
+      viceHead: {
+        name: "Michael Brown",
+        email: "michael@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/michaelb"
+      },
+      teamMembers: [
+        {
+          name: "Test member 2",
+          email: "",
+          linkedin: "",
+          github: ""
+        },
+        {
+          name: "Testmember 2",
+          email: "alice@systemvetardagen.se",
+          github: "https://github.com/alicej"
+        }
+      ]
+    },
+    "Business Relations and Sales": {
+      head: {
+        name: "Sarah Connor",
+        email: "sarah@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/sarahc"
+      },
+      viceHead: {
+        name: "James Wilson",
+        email: "james@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/jamesw"
+      },
+      teamMembers: [
+        {
+          name: "Test member 2",
+          email: "",
+          linkedin: "",
+          github: ""
+        },
+        {
+          name: "Testmember 2",
+          email: "alice@systemvetardagen.se",
+          github: "https://github.com/alicej"
+        }
+      ]
+    },
+    "Logistics": {
+      head: {
+        name: "David Miller",
+        email: "david@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/davidm"
+      },
+      viceHead: {
+        name: "Lisa Anderson",
+        email: "lisa@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/lisaa"
+      },
+      teamMembers: [
+        {
+          name: "Test member 2",
+          email: "",
+          linkedin: "",
+          github: ""
+        },
+        {
+          name: "Testmember 2",
+          email: "alice@systemvetardagen.se",
+          github: "https://github.com/alicej"
+        }
+      ]
+    },
+    "Project Management": {
+      head: {
+        name: "Robert Taylor", 
+        email: "robert@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/robertt"
+      },
+      viceHead: {
+        name: "Emily Davis",
+        email: "emily@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/emilyd"
+      },
+      teamMembers: [
+        {
+          name: "Test member 2",
+          email: "",
+          linkedin: "",
+          github: ""
+        },
+        {
+          name: "Testmember 2",
+          email: "alice@systemvetardagen.se",
+          github: "https://github.com/alicej"
+        }
+      ]
+    },
+    "Events": {
+      head: {
+        name: "Sophie Martin",
+        email: "sophie@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/sophiem"
+      },
+      viceHead: {
+        name: "Daniel Lee",
+        email: "daniel@systemvetardagen.se",
+        linkedin: "https://linkedin.com/in/daniell"
+      },
+      teamMembers: [
+        {
+          name: "Test member 2",
+          email: "",
+          linkedin: "",
+          github: ""
+        },
+        {
+          name: "Testmember 2",
+          email: "alice@systemvetardagen.se",
+          github: "https://github.com/alicej"
+        }
+      ]
+    }
+  };
